+++ packages/utils/sets-utils/src/expr-utils.test.js
@@ -2,15 +2,11 @@ import { describe, it, expect } from 'vitest';
 import {
   stratifyExpressionData,
   aggregateStratifiedExpressionData,
-} from '@vitessce/sets-utils';
-import {
-  summarizeStratifiedExpressionData,
-  histogramStratifiedExpressionData,
-} from './expr-hooks.js';
+} from './expr-utils.js';
 
-describe('Utility functions for processing expression data for statistical plots', () => {
-  describe('summarizeStratifiedExpressionData function', () => {
-    it('computes summarized information accurately', () => {
+describe('Utility functions for processing expression data', () => {
+  describe('stratifyExpressionData function', () => {
+    it('stratify by cell set, then sample set', () => {
       const sampleEdges = new Map([
         ['cell1-1', 'donor1'],
         ['cell1-2', 'donor1'],
@@ -85,32 +81,18 @@ describe('Utility functions for processing expression data for statistical plots
         geneSelection, cellSetSelection, cellSetColor,
         featureValueTransform, featureValueTransformCoefficient,
       );
+
+      expect([...result.keys()]).toEqual([['Cell type', 'T cell'], ['Cell type', 'B cell']]);
+      expect([...result.get(['Cell type', 'T cell']).keys()]).toEqual([['Clinical groups', 'AKI'], ['Clinical groups', 'CKD']]);
+      expect(result.get(['Cell type', 'T cell']).get(['Clinical groups', 'AKI']).get('Gene 1').length).toBe(2);
+      expect(result.get(['Cell type', 'T cell']).get(['Clinical groups', 'AKI']).get('Gene 1')).toEqual([10, 30]);
+      expect(exprMax).toEqual(41);
+
       const aggregateData = aggregateStratifiedExpressionData(
         result, geneSelection,
       );
-      const summaryResult = summarizeStratifiedExpressionData(aggregateData, true);
 
-      expect([...summaryResult.keys()]).toEqual([['Cell type', 'T cell'], ['Cell type', 'B cell']]);
-      expect([...summaryResult.get(['Cell type', 'T cell']).keys()]).toEqual([['Clinical groups', 'AKI'], ['Clinical groups', 'CKD']]);
-      expect(Object.keys(summaryResult.get(['Cell type', 'T cell']).get(['Clinical groups', 'AKI']))).toEqual([
-        'quartiles',
-        'range',
-        'whiskers',
-        'chauvenetRange',
-        'nonOutliers',
-      ]);
-
-      const histogramResult = histogramStratifiedExpressionData(summaryResult, 16, null);
-
-      expect(Object.keys(histogramResult)).toEqual([
-        'groupSummaries',
-        'groupData',
-        'groupBins',
-        'groupBinsMax',
-        'y',
-      ]);
-      expect(histogramResult.groupSummaries.map(d => d.key)).toEqual([['Cell type', 'T cell'], ['Cell type', 'B cell']]);
-      expect(histogramResult.groupBinsMax).toEqual(1);
+      // TODO: add expect statements that check the aggregateData result.
     });
   });
 });
